requirements.txt
# CA-LLE package — Python dependencies
# Recommended: install from the repository root with `pip install -r requirements.txt`
# so PyTorch CUDA wheels match your driver. This file mirrors the root list for
# convenience (e.g. browsing only the `ca_lle/` subtree).

# PyTorch: install a CUDA build matching your system from https://pytorch.org
torch>=2.0.0
torchvision>=0.15.0

numpy>=1.22.0
Pillow>=9.0.0
tqdm>=4.64.0
opencv-python>=4.7.0

open-clip-torch>=2.20.0
lpips>=0.1.4
pytorch-msssim>=1.0.0
scikit-image>=0.20.0
